# Reduced memory usage with a pre-generated index