--- a/src/index.hpp
+++ b/src/index.hpp
@@ -168,7 +168,7 @@ private:
     const References& references;
     kmer_lookup mers_index; // k-mer -> (offset in flat_vector, occurence count )
 
-    void index_vector(const hash_vector& h_vector, float f);
+    void index_vector(const ind_mers_vector &mers, float f);
     ind_mers_vector generate_and_sort_seeds() const;
 };