--- a/src/randstrobes.hpp
+++ b/src/randstrobes.hpp
@@ -8,9 +8,11 @@
 
 // only used during index generation
 struct MersIndexEntry {
+
+    using packed_t = uint32_t;
     uint64_t hash;
     uint32_t position;
-    int32_t packed; // packed representation of ref_index and strobe offset
+    packed_t packed; // packed representation of ref_index and strobe offset
 
     bool operator< (const MersIndexEntry& other) const {
         return std::tie(hash, position, packed) < std::tie(other.hash, other.position, other.packed);