Here’s a draft proposal for reducing memory requirements.

As discussed in #173, each entry in the index hash table (`mers_index`) maps a randstrobe to an (offset, count) pair that point to a slice of the flat vector, where the actual information about the randstrobe is stored. For CHM13, ~500 million unique randstrobes are generated, so the hash table has 500 million entries. However, 97.4% of the randstrobes occur only once, so to store one of those, we need an (offset, count) pair (8 bytes) and the `ReferenceMer` entry in the flat vector itself (another 8 bytes). The idea implemented here is to store those randstrobes directly in the hash table:
- The most significant bit of the "count" value is used as a marker
- If it is set, the (offset, count) pair is re-interpreted as a `ReferenceMer`, where offset is interpreted as position and count is interpreted as the "packed" reference index and strobemer offset.
- If the bit is not set, (offset, count) are interpreted as before.

On constructing the index, we only add randstrobes with count greater than one to the flat vector, which gives us the memory savings.

The reason why I’ve marked this as "draft" is that peak memory on index construction is a bit larger because I had to undo the optimization of throwing away the hash values and filling the mers_index afterwards because I now need both to be available. One option would be to fall back to the previous way of doing things if `--use-index` is not provided.

I also have been working on a way to reduce the memory usage even during index creation time, but that is in a different branch

Here are some numbers on CHM13 using the 10 M 200 bp paired-end reads from your evaluation dataset.

| | main | this branch |
|-|-|-|
| Size of `.sti` file | 11.7 GiB | 8.2 GiB |
| Peak mem when creating index |28.6 GiB | 29.6 GiB |
| Time to create index | ~5:00 | ~5:00 |
| Peak mem with `-t 8 --use-index` |24.4 GiB | 20.7 GiB |
| (User) runtime mapping 10 M paired-end reads | ~2500s | ~2500s |

~~I think there is something fishy going on with the huge peak memory usage while mapping reads. I only discovered this just now when I wanted to see whether runtime changes when mapping a large number of reads. That is a separate issue however.~~ Fixed by #181. (*Edit:* Changed table above to no longer show peak memory for mapping 10 million reads. It’s the same as when mapping any number of reads. *main* includes the fix in #181.)

Note also how the index itself only contains 8.2 GiB of data. When it’s loaded, that becomes 17 GiB because of the maximum load factor of the hashtable and because its size can only be a power of two. If we could reduce this to the absolute minimum somehow, strobealign would only need ~12 GiB of RAM (3.2 GiB for the reference plus 8.2 GiB for the index plus a little extra).