@ksahlin 2022-12-16
Also, FWIW, ~12 GiB looks suspiciously close to minimap2's peak memory usage. Perhaps Heng has though exactly about this in his hash table implementation.