@marcelm 2022-12-16
>     1. Wouldn't this also give us a tad speedup if reads mostly have unique seeds? (since we do not have to access the flat_vector as much?

If there is a speedup, it’s not much. The runtime was a bit lower (maybe 1-2%), but I didn’t run it often enough so that I could say anything with statistical certainty.

>     2. Wouldn't the following reduce peak memory during indexing?
>        A. construct and sort `flat_vector`
>        B. save `flat_vector` to disk
>        C. Read `flat_vector`as a stream and construct the hash table (using your new approach of storing randstorbes directly in hash table). Keeping offset incrementing only for non-unique entries. (here things over hard cutoff of 1000 can also be ignored)
>        D. Read in the `flat_vector` again (only randstrobes occurring more than once).
> 
> It would only increase indexing runtime slightly I think. While keeping peak memory as low as for alignment.

That’s a good suggestion!

Maybe I should mention the other thing I have actually already implemented (in branch `index`). It doesn’t quite work (yet) as I want it, but I want to mention the idea: The reason the flat vector is created in full in memory at all is that this is needed for sorting it. But we no longer need the hashes with count==1 to be sorted. (And technically, we don’t even need the hashes with count > 1 to be sorted, the ones with the same hash value just need to be adjacent to each other so that they can be looked up together.)

So we can just iterate over all randstrobes and directly insert them into the hash table and the flat vector. More precisely:
- Check if the randstrobe is already in the table. If not, add it as a "direct" entry.
- If it is, convert the entry to an "indirect" one if count=1 (extract the existing entry, add it to the flat vector), then add the new randstrobe.
- At the end, sort the flat vector, and go over its entries to fix up the "offset" fields in the hash table.
(This requires that the flat vector still contains the hashes.)
The issue I have at the moment is that I need to iterate over the all randstrobes once in the beginning to count the unique hashes so that I can reserve the correct size. If I don’t do that, the hash table gets resized when adding entries, which blows up memory usage and runtime. 

>     3. I am also interested in what you wrote about the hash table. Potentially a big gain to be made! Did you try if the other implementation in robin_hood is better? It should be more memory sparse, but not near the minimum I predict.

I tried it just now. Peak memory usage with `unordered_node_map` is 24 GB.