@ksahlin 2022-12-17
> That’s a good suggestion! Maybe I should mention the other thing I have actually already implemented (in branch index). 

When I suggest something, you typically have something better going on ^^. Do you have any estimate on how much the peak memory for this approach would be? Would be great have a peak memory that is not much more than during alignment. Maybe wishful thinking. 

Also, to challenge your approach with 'extreme scenarios', the approach would not have as much gain over our previous approach for very repetitive genomes withour many unique seeds. However, CHM13 is a good standard. Just pointing it out for extreme cases such as repetitive plant genomes. Good that we have those datasets to benchmark on.

> The issue I have at the moment is that I need to iterate over the all randstrobes once in the beginning to count the unique hashes so that I can reserve the correct size. If I don’t do that, the hash table gets resized when adding entries, which blows up memory usage and runtime.

Indexing time is not crucial imo. Better to have a stable peak memory. Also, since hash table look ups are not a major part of the runtime, maybe it is worth looking at other header-only memory sparse implementations, for example [sparsepp](https://github.com/greg7mdp/sparsepp), or find any other alternative from [this benchmark](https://martin.ankerl.com/2022/08/27/hashmap-bench-01/#memory-usage)? [This](https://github.com/lh3/minimap2/blob/master/khash.h) seems to be minimap2's hash table implementation. Something about quadratic probing. Haven't looked up what it means yet. 

Another direction is to use 32bit hashes that I played around with a bit before and suggested in https://github.com/ksahlin/strobealign/issues/27. My main concern with the 32bit int hash table are collisions for very large genomes as we are capped by 4.2billion entries.

Just brainstorming a bit. Of course, the trade-off with how much slower it gets should IMO always be consulted. 


> I tried it just now. Peak memory usage with unordered_node_map is 24 GB.

Okay, that wasn't a viable direction.