@ksahlin 2022-12-18
I was curious so I checked the results for `rye` (7.3G) and `maize` (2.4G) genomes with seeds from 2*150nt reads.
For rye, 83% of the seeds occur once out of 668M total distinct hash values. Maize has about 86% of the seeds occurring only once from a total of 208M distinct hash values. So still big gain to be made with your approach.