@marcelm 2022-12-19
Nice results for rye and maize! 

> Do you have any estimate on how much the peak memory for this approach would be? Would be great have a peak memory that is not much more than during alignment. Maybe wishful thinking.

It would indeed be nearly the same as during alignment. I suggest that I open a separate PR for this. I think it is still valuable to first get this PR to a stage where it can be merged because that other method is a bit slower, so maybe it’s worth to have both and then be able to switch between them (high memory/fast vs low memory/slow).

> Indexing time is not crucial imo. Better to have a stable peak memory. Also, since hash table look ups are not a major part of the runtime, maybe it is worth looking at other header-only memory sparse implementations, for example [sparsepp](https://github.com/greg7mdp/sparsepp), or find any other alternative from [this benchmark](https://martin.ankerl.com/2022/08/27/hashmap-bench-01/#memory-usage)?

I tested gtl::flat_hash_map a couple of days ago. It used more memory and was a lot slower when creating the index. I’ll need to test sparsepp, but it claims to also have one byte overhead per entry, so this would be the same as what we have now. I haven’t systematically tried the other hash table implementations.

> [This](https://github.com/lh3/minimap2/blob/master/khash.h) seems to be minimap2's hash table implementation. Something about quadratic probing. Haven't looked up what it means yet.

To try this, I think we would need to find a C++ wrapper for it (haven’t searched, this probably exists somewhere).

> Another direction is to use 32bit hashes that I played around with a bit before and suggested in #27. My main concern with the 32bit int hash table are collisions for very large genomes as we are capped by 4.2billion entries.

I’ll comment over at #27.